transport: Refactor 'Listen' code
At the moment, the Listen code is implemented for each platform, but the
URL parsing code, and `unix` and `tcp` code are the same.
There's also no 'generic' fallback for other platforms which don't need
a vsock implementation.

This commit replaces the per-platform Listen implementation with
a generic one doing the URL parsing and then calling a per-platform
`listenURL`. It also introduces a `defaultListenURL` function with the
`unix` and `tcp` code paths.
This way, the per-platform files only contain the platform-specific
code.
