+++ pkg/transport/listen_darwin.go
@@ -12,11 +12,7 @@ import (
 
 const DefaultURL = "vsock://null:1024/vm_directory"
 
-func Listen(endpoint string) (net.Listener, error) {
-	parsed, err := url.Parse(endpoint)
-	if err != nil {
-		return nil, err
-	}
+func listenURL(parsed *url.URL) (net.Listener, error) {
 	switch parsed.Scheme {
 	case "vsock":
 		port, err := strconv.Atoi(parsed.Port())
@@ -31,12 +27,8 @@ func Listen(endpoint string) (net.Listener, error) {
 			Name: path,
 			Net:  "unix",
 		})
-	case "unix":
-		return net.Listen("unix", parsed.Path)
-	case "tcp":
-		return net.Listen("tcp", parsed.Host)
 	default:
-		return nil, errors.New("unexpected scheme")
+		return defaultListenURL(parsed)
 	}
 }
