+++ pkg/transport/listen_linux.go
@@ -1,7 +1,6 @@
 package transport
 
 import (
-	"errors"
 	"net"
 	"net/url"
 	"strconv"
@@ -11,11 +10,7 @@ import (
 
 const DefaultURL = "vsock://:1024"
 
-func Listen(endpoint string) (net.Listener, error) {
-	parsed, err := url.Parse(endpoint)
-	if err != nil {
-		return nil, err
-	}
+func listenURL(parsed *url.URL) (net.Listener, error) {
 	switch parsed.Scheme {
 	case "vsock":
 		port, err := strconv.Atoi(parsed.Port())
@@ -23,11 +18,9 @@ func Listen(endpoint string) (net.Listener, error) {
 			return nil, err
 		}
 		return mdlayhervsock.Listen(uint32(port), nil)
-	case "unix", "unixpacket":
+	case "unixpacket":
 		return net.Listen(parsed.Scheme, parsed.Path)
-	case "tcp":
-		return net.Listen("tcp", parsed.Host)
 	default:
-		return nil, errors.New("unexpected scheme")
+		return defaultListenURL(parsed)
 	}
 }
