+++ pkg/transport/listen_windows.go
@@ -1,7 +1,6 @@
 package transport
 
 import (
-	"errors"
 	"net"
 	"net/url"
 
@@ -10,11 +9,7 @@ import (
 
 const DefaultURL = "vsock://00000400-FACB-11E6-BD58-64006A7986D3"
 
-func Listen(endpoint string) (net.Listener, error) {
-	parsed, err := url.Parse(endpoint)
-	if err != nil {
-		return nil, err
-	}
+func listenURL(parsed *url.URL) (net.Listener, error) {
 	switch parsed.Scheme {
 	case "vsock":
 		svcid, err := hvsock.GUIDFromString(parsed.Hostname())
@@ -25,11 +20,7 @@ func Listen(endpoint string) (net.Listener, error) {
 			VMID:      hvsock.GUIDWildcard,
 			ServiceID: svcid,
 		})
-	case "unix":
-		return net.Listen(parsed.Scheme, parsed.Path)
-	case "tcp":
-		return net.Listen("tcp", parsed.Host)
 	default:
-		return nil, errors.New("unexpected scheme")
+		return defaultListenURL(parsed)
 	}
 }
